refatoração
Alteração de Integer para Long;
remoçõ

+++ src/main/java/br/com/wti/erp/repository/MonitorRepository.java
@@ -54,7 +54,7 @@ public class MonitorRepository implements Serializable {
 		}
 	}
 	
-	public Long getQuantityForProject(Integer project) {
+	public Long getQuantityForProject(Long project) {
 		try {
 			String jpql = "SELECT COUNT(m) "
 					+ "FROM Monitor m "
